+++ Assets/Scripts/ActionsController.cs
@@ -34,6 +34,7 @@ namespace EVRC
             // Menu/UI buttons
             MenuBack,
             MenuSelect,
+            MenuNestedToggle,
             MenuNavigateUp,
             MenuNavigateDown,
             MenuNavigateLeft,
@@ -53,6 +54,12 @@ namespace EVRC
             MenuNavigateTrackpad,
             UINavigateTrackpad,
             UITabTrackpad,
+            // Joystick POV/Menu/UI
+            POV1Joystick,
+            POV2Joystick,
+            MenuNavigateJoystick,
+            UINavigateJoystick,
+            UITabJoystick,
         }
 
         public enum OutputAction
@@ -72,6 +79,7 @@ namespace EVRC
             // Menu
             MenuBack,
             MenuSelect,
+            MenuNestedToggle,
             MenuNavigate,
             // UI Navigate/Tab
             UIBack,
@@ -148,6 +156,7 @@ namespace EVRC
         private delegate void Vector2InputActionHandler(Hand hand, Vector2 newState);
         private delegate IEnumerator TrackpadInputActionHandler(InputAction inputAction, Hand hand, DynamicRef<Vector2> position, Ref<bool> running);
         private delegate void TrackpadPressActionHandler(InputAction inputAction, Hand hand, Vector2 position, bool newState);
+        private delegate void JoystickPositionChangeActionHandler(InputAction inputAction, Hand hand, Vector2 axis);
 
         private delegate bool EmitStateChangeDelegate(bool newState);
         private delegate bool EmitDirectionStateChangeDelegate(Direction direction, bool newState);
@@ -156,13 +165,16 @@ namespace EVRC
         private Dictionary<InputAction, Vector2InputActionHandler> vector2InputActionHandlers;
         private Dictionary<InputAction, TrackpadInputActionHandler> trackpadInputActionHandlers;
         private Dictionary<InputAction, TrackpadPressActionHandler> trackpadPressActionHandlers;
+        private Dictionary<InputAction, JoystickPositionChangeActionHandler> joystickActionHandlers;
 
         private Dictionary<InputAction, OutputAction> simpleBooleanActionMapping;
         private Dictionary<InputAction, (OutputAction, Direction)> directionalBooleanActionMapping;
         private Dictionary<InputAction, OutputAction> directionalTrackpadSlideMapping;
         private Dictionary<InputAction, (OutputAction, OutputAction)> trackpadPressActionMapping;
+        private Dictionary<InputAction, OutputAction> joystickDirectionActionMapping;
 
         private Dictionary<(InputAction, Hand), Action> trackpadPressUnpressHandler = new Dictionary<(InputAction, Hand), Action>();
+        private Dictionary<(InputAction, Hand), (Direction, Action)> joystickDirectionUnpressHandler = new Dictionary<(InputAction, Hand), (Direction, Action)>();
 
         public static string[] GetBindingNames(IBindingsController bindingsController, OutputAction outputAction, NameType nameType)
         {
@@ -176,8 +188,8 @@ namespace EVRC
                 case OutputAction.ButtonPrimary: return MergeBindings(InputAction.ButtonPrimary);
                 case OutputAction.ButtonSecondary: return MergeBindings(InputAction.ButtonSecondary);
                 case OutputAction.ButtonAlt: return MergeBindings(InputAction.ButtonAlt);
-                case OutputAction.POV1: return MergeBindings(InputAction.ButtonPOV1, InputAction.POV1Trackpad);
-                case OutputAction.POV2: return MergeBindings(InputAction.ButtonPOV2, InputAction.POV2Trackpad);
+                case OutputAction.POV1: return MergeBindings(InputAction.ButtonPOV1, InputAction.POV1Trackpad, InputAction.POV1Joystick);
+                case OutputAction.POV2: return MergeBindings(InputAction.ButtonPOV2, InputAction.POV2Trackpad, InputAction.POV2Joystick);
             }
 
             throw new Exception(string.Format("OutputAction.{0} is not handled by GetBindingNames", outputAction));
@@ -192,10 +204,12 @@ namespace EVRC
             vector2InputActionHandlers = new Dictionary<InputAction, Vector2InputActionHandler> { };
             trackpadInputActionHandlers = new Dictionary<InputAction, TrackpadInputActionHandler> { };
             trackpadPressActionHandlers = new Dictionary<InputAction, TrackpadPressActionHandler> { };
+            joystickActionHandlers = new Dictionary<InputAction, JoystickPositionChangeActionHandler>();
             simpleBooleanActionMapping = new Dictionary<InputAction, OutputAction>();
             directionalBooleanActionMapping = new Dictionary<InputAction, (OutputAction, Direction)>();
             directionalTrackpadSlideMapping = new Dictionary<InputAction, OutputAction>();
             trackpadPressActionMapping = new Dictionary<InputAction, (OutputAction, OutputAction)>();
+            joystickDirectionActionMapping = new Dictionary<InputAction, OutputAction>();
 
             // Basic interactions
             MapBooleanInputActionToOutputAction(InputAction.InteractUI, OutputAction.InteractUI);
@@ -213,6 +227,7 @@ namespace EVRC
             // Menu/UI buttons
             MapBooleanInputActionToOutputAction(InputAction.MenuBack, OutputAction.MenuBack);
             MapBooleanInputActionToOutputAction(InputAction.MenuSelect, OutputAction.MenuSelect);
+            MapBooleanInputActionToOutputAction(InputAction.MenuNestedToggle, OutputAction.MenuNestedToggle);
             MapBooleanInputActionToDirectionOutputAction(InputAction.MenuNavigateUp, OutputAction.MenuNavigate, Direction.Up);
             MapBooleanInputActionToDirectionOutputAction(InputAction.MenuNavigateDown, OutputAction.MenuNavigate, Direction.Down);
             MapBooleanInputActionToDirectionOutputAction(InputAction.MenuNavigateLeft, OutputAction.MenuNavigate, Direction.Left);
@@ -236,12 +251,19 @@ namespace EVRC
             MapTrackpadPressToDirectionAndButtonOptionAction(InputAction.UINavigateTrackpad, OutputAction.UINavigate, OutputAction.UISelect);
             trackpadInputActionHandlers[InputAction.UITabTrackpad] = OnUITabTrackpadInput;
             trackpadPressActionHandlers[InputAction.UITabTrackpad] = OnUITabTrackpadPress;
+            // Joystick POV/Menu/UI
+            MapJoystickToDirectionOutputAction(InputAction.POV1Joystick, OutputAction.POV1);
+            MapJoystickToDirectionOutputAction(InputAction.POV2Joystick, OutputAction.POV2);
+            MapJoystickToDirectionOutputAction(InputAction.MenuNavigateJoystick, OutputAction.MenuNavigate);
+            MapJoystickToDirectionOutputAction(InputAction.UINavigateJoystick, OutputAction.UINavigate);
+            joystickActionHandlers[InputAction.UITabJoystick] = OnUITabJoystickAxisChange;
         }
 
         void OnDisable()
         {
         }
 
+        #region Binding implementation interface
         /**
          * Interface for binding implementations to update the pose of a hand
          */
@@ -296,6 +318,22 @@ namespace EVRC
             }
         }
 
+        /**
+         * Interface for input binding implementations to call when a joystick's axis has changed
+         */
+        public void TriggerJoystickAxisChangeAction(InputAction inputAction, Hand hand, Vector2 axis)
+        {
+            if (joystickActionHandlers.ContainsKey(inputAction))
+            {
+                joystickActionHandlers[inputAction](inputAction, hand, axis);
+            }
+            else
+            {
+                Debug.LogWarningFormat("No joystick handler for input action: {0}", inputAction.ToString());
+            }
+        }
+        #endregion
+
         private void EmitActionStateChange(Hand hand, OutputAction action, bool state)
         {
             var ev = new ActionChange(hand, action, state);
@@ -322,6 +360,7 @@ namespace EVRC
             }
         }
 
+        #region Boolean
         private void MapBooleanInputActionToOutputAction(InputAction inputAction, OutputAction outputAction)
         {
             simpleBooleanActionMapping[inputAction] = outputAction;
@@ -358,7 +397,9 @@ namespace EVRC
                 Debug.LogWarningFormat("No directional boolean action mapping for input action: {0}", inputAction.ToString());
             }
         }
+        #endregion
 
+        #region Trackpad slide
         private void MapTrackpadSlideToDirectionOutputAction(InputAction inputAction, OutputAction outputAction)
         {
             directionalTrackpadSlideMapping[inputAction] = outputAction;
@@ -403,6 +444,40 @@ namespace EVRC
                 });
         }
 
+        public class TrackpadInterval
+        {
+            // A "safe" fallback interval that allows about 1 interval for a swipe across most of most trackpads
+            public static readonly TrackpadInterval Default = Circular(.75f);
+
+            public static TrackpadInterval Circular(float interval)
+                => new TrackpadInterval { Horizontal = interval, Vertical = interval };
+            public static TrackpadInterval Oval(float horizontal, float vertical)
+                => new TrackpadInterval { Horizontal = horizontal, Vertical = vertical };
+
+            public float Horizontal { get; private set; }
+            public float Vertical { get; private set; }
+
+            protected TrackpadInterval() { }
+
+            /**
+             * Get the Horizontal or Vertical interval as a float depending on the given direction
+             */
+            public float ForDirection(Direction dir)
+            {
+                switch (dir)
+                {
+                    case Direction.Up:
+                    case Direction.Down:
+                        return Vertical;
+                    case Direction.Right:
+                    case Direction.Left:
+                        return Horizontal;
+                    default:
+                        throw new ArgumentException("Unknown direction");
+                }
+            }
+        }
+
         private IEnumerator DoAbstractTrackpadInput(
             InputAction inputAction, Hand hand, DynamicRef<Vector2> position, Ref<bool> running,
             EmitDirectionStateChangeDelegate emitDirectionStateChange
@@ -411,12 +486,7 @@ namespace EVRC
             // Wait a tick before starting, a race condition results in the current position always starting as (0, 0)
             yield return null;
 
-            var trackpadInterval = ActionsControllerBindingsLoader.CurrentBindingsController?.GetTrackpadSwipeInterval(hand) ?? 0;
-            if (trackpadInterval == 0)
-            {
-                // When trackpad is unknown fallback to a "safe" interval that allows about 1 interval for a swipe across most of the pad
-                trackpadInterval = 1f;
-            }
+            var trackpadInterval = ActionsControllerBindingsLoader.CurrentBindingsController?.GetTrackpadSwipeInterval(hand);
 
             Vector2 anchorPos = position.Current;
             yield return null;
@@ -426,12 +496,11 @@ namespace EVRC
                 var deltaPos = pos - anchorPos;
                 float magnitude = 0;
                 Direction dir = GetLargestVectorDirection(deltaPos, ref magnitude);
-                if (magnitude >= trackpadInterval)
+                if (magnitude >= trackpadInterval.ForDirection(dir))
                 {
                     anchorPos = pos;
                     if (emitDirectionStateChange(dir, true))
                     {
-
                         // Wait long enough for ED to recieve any keypresses
                         yield return KeyboardInterface.WaitForKeySent();
 
@@ -442,7 +511,9 @@ namespace EVRC
                 yield return null;
             }
         }
+        #endregion
 
+        #region Trackpad press
         private void MapTrackpadPressToDirectionAndButtonOptionAction(InputAction inputAction, OutputAction directionOutputAction, OutputAction centerOutputAction)
         {
             trackpadPressActionMapping[inputAction] = (directionOutputAction, centerOutputAction);
@@ -537,7 +608,93 @@ namespace EVRC
                 }
             }
         }
+        #endregion
 
+        #region Joystick
+        private void MapJoystickToDirectionOutputAction(InputAction inputAction, OutputAction outputAction)
+        {
+            joystickActionHandlers[inputAction] = OnMappedJoystickAxisChange;
+            joystickDirectionActionMapping[inputAction] = outputAction;
+        }
+
+        private void OnMappedJoystickAxisChange(InputAction inputAction, Hand hand, Vector2 axis)
+        {
+            if (!joystickDirectionActionMapping.ContainsKey(inputAction))
+            {
+                Debug.LogWarningFormat("No joystick input action mapping for input action: {0}", inputAction.ToString());
+                return;
+            }
+            var outputAction = joystickDirectionActionMapping[inputAction];
+
+            DoAbstractJoystickInput(
+                inputAction, hand, axis,
+                (dir, newButtonState) =>
+                {
+                    EmitDirectionActionStateChange(hand, outputAction, dir, newButtonState);
+                    return true;
+                });
+        }
+
+        private void OnUITabJoystickAxisChange(InputAction inputAction, Hand hand, Vector2 axis)
+        {
+            DoAbstractJoystickInput(
+                inputAction, hand, axis,
+                (dir, newButtonState) =>
+                {
+                    switch (dir)
+                    {
+                        case Direction.Left:
+                            EmitActionStateChange(hand, OutputAction.UITabPrevious, newButtonState);
+                            return true;
+                        case Direction.Right:
+                            EmitActionStateChange(hand, OutputAction.UITabNext, newButtonState);
+                            return true;
+                        default:
+                            return false;
+                    }
+                });
+        }
+
+        private void DoAbstractJoystickInput(
+            InputAction inputAction, Hand hand, Vector2 axis,
+            EmitDirectionStateChangeDelegate emitDirectionStateChange
+        )
+        {
+            bool isReleased = Mathf.Abs(axis.x) < float.Epsilon && Mathf.Abs(axis.y) < float.Epsilon;
+
+            Direction? dir = null;
+            if (!isReleased)
+            {
+                float _magnitude = 0;
+                dir = GetLargestVectorDirection(axis, ref _magnitude);
+            }
+
+            // Release any previous press whether we have an axis=0,0 release, or rotated to a different direction
+            if (joystickDirectionUnpressHandler.ContainsKey((inputAction, hand)))
+            {
+                (Direction prevDirection, Action unpressHandler) = joystickDirectionUnpressHandler[(inputAction, hand)];
+                if (dir == null || dir != prevDirection)
+                {
+                    unpressHandler();
+                }
+            }
+
+            if (dir != null)
+            {
+                if (emitDirectionStateChange(dir.Value, true))
+                {
+                    joystickDirectionUnpressHandler[(inputAction, hand)] = (dir.Value, () =>
+                    {
+                        joystickDirectionUnpressHandler.Remove((inputAction, hand));
+                        emitDirectionStateChange(dir.Value, false);
+                    }
+                    );
+                }
+            }
+        }
+        #endregion
+
+        #region Reset Seated Position
         private readonly HashSet<Hand> maybeResetSeatedPositionHandPressed = new HashSet<Hand>();
         private bool maybeResetSeatedPositionBothHandsPressed = false;
 
@@ -552,6 +709,7 @@ namespace EVRC
                 maybeResetSeatedPositionBothHandsPressed = bothPressed;
             }
         }
+        #endregion
 
         private Direction GetLargestVectorDirection(Vector2 v, ref float magnitude)
         {